// components/types.ts
export type LocationData = {
    title: string;
    description: string;
    image: string;
    packagePrice: string;
    duration: string;
    bestTime: string;
    trekDistance: string;
    maxAltitude: string;
    imageTag:string,
    imagetag2:string,
    imagetag3:string,
    Weekends:string,
    overview: string[];
    Payment:string;
    Longest:string;
    itinerary: {
      day: number;
      description: string;
      image: string;
      Topic:string;
    }[];
  };
  